feat(cli): don't print ignored vulnerabilities
